+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Experimental race import for God Mode Tier patreons to trial.
 * Refactored existing item search to improve lookups for future improvements.
 * Homebrew monster import toggle.
+* Module setting for monsters have vision toggle.
 
 # v 0.2.0